Move functions back to lib/
