+++ lib/handleExternalLinks.js
@@ -1,8 +1,10 @@
-import { WebBrowser } from 'expo';
+import * as WebBrowser from "expo-web-browser";
 
-export default function(webview, url) {
+const handleExternalLinks = async (url, webViewRef) => {
   if (!/timeoverflow/.test(url)) {
-    webview.goBack();
-    WebBrowser.openBrowserAsync(url);
+    webViewRef.current?.goBack();
+    await WebBrowser.openBrowserAsync(url);
   }
-}
+};
+
+export default handleExternalLinks;
