+++ lib/injectCustomJavaScript.js
@@ -1,19 +1,19 @@
-import { registerForPushNotificationsAsync } from './pushNotifications';
-
-const handleLoggedInPage = async (webview) => {
-  const token = await registerForPushNotificationsAsync();
-
-  webview.injectJavaScript(`window.TimeOverflowRegisterExpoDeviceToken(\'${token}\');`);
+const handleLoggedInPage = (webview, token) => {
+  webview?.injectJavaScript(
+    `window.TimeOverflowRegisterExpoDeviceToken(\'${token}\');`
+  );
 };
 
 const handleLoginPage = (webview) => {
-  webview.injectJavaScript("$('#user_remember_me').prop('checked', true);")
+  webview?.injectJavaScript("$('#user_remember_me').prop('checked', true);");
 };
 
-export default function(webview, url) {
+const injectCustomJavaScript = (webViewRef, token, url) => {
   if (/members/.test(url)) {
-    handleLoggedInPage(webview);
+    handleLoggedInPage(webViewRef.current, token);
   } else if (/sign_in|login/.test(url)) {
-    handleLoginPage(webview);
+    handleLoginPage(webViewRef.current);
   }
-}
+};
+
+export default injectCustomJavaScript;
